StudderServer/requirements.txt
django >=1.9, <2
django
djangorestframework
mysqlclient
mysql-connector-python